[#2] chown : helper 상수 데이터를 constants로 이동
